+++ src/Symfony/Component/Ldap/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 -----
 
  * added `EntryManager::move`, not implementing it is deprecated
+ * Added pagination support to the ExtLdap adapter with the pageSize query option
 
 4.2.0
 -----